Move control structure tests to a dedicated place.
