+++ joern-cli/frontends/rust2cpg/src/test/scala/io/joern/rust2cpg/passes/ast/SimpleAstCreationPassTest.scala
@@ -170,151 +170,6 @@ class SimpleAstCreationPassTest extends Rust2CpgSuite {
 
   //
 
-  "test 12" in {
-    val cpg = code("""
-        |fn main(x: i32, y: i32) {
-        | if x > y {
-        |  foo();
-        | }
-        |}
-        |""".stripMargin)
-
-    inside(cpg.method.name("main").controlStructure.controlStructureTypeExact(ControlStructureTypes.IF).l) {
-      case (ifExpr: ControlStructure) :: Nil =>
-        ifExpr.code shouldBe "if x > y {\n  foo();\n }"
-
-        inside(ifExpr.condition.isCall.l) { case (condition: Call) :: Nil =>
-          condition.code shouldBe "x > y"
-          condition.name shouldBe Operators.greaterThan
-          condition.methodFullName shouldBe Operators.greaterThan
-
-          inside(condition.argument.l) { case (lhs: Identifier) :: (rhs: Identifier) :: Nil =>
-            lhs.name shouldBe "x"
-            rhs.name shouldBe "y"
-          }
-        }
-
-        inside(ifExpr.whenTrue.isBlock.l) { case (thenBlock: Block) :: Nil =>
-          thenBlock.astChildren.isCall.name.l shouldBe List("foo")
-        }
-
-        ifExpr.whenFalse.l shouldBe empty
-    }
-  }
-
-  "test 13" in {
-    val cpg = code("""
-        |fn main(x: i32, y: i32) {
-        | if x == y {
-        |  foo();
-        | } else {
-        |  bar();
-        | }
-        |}
-        |""".stripMargin)
-
-    inside(cpg.method.name("main").controlStructure.controlStructureTypeExact(ControlStructureTypes.IF).l) {
-      case (ifExpr: ControlStructure) :: Nil =>
-        inside(ifExpr.condition.isCall.l) { case (condition: Call) :: Nil =>
-          condition.code shouldBe "x == y"
-          condition.name shouldBe Operators.equals
-          condition.methodFullName shouldBe Operators.equals
-        }
-
-        ifExpr.whenTrue.isBlock.astChildren.isCall.name.l shouldBe List("foo")
-
-        inside(ifExpr.whenFalse.isControlStructure.controlStructureTypeExact(ControlStructureTypes.ELSE).l) {
-          case (elseExpr: ControlStructure) :: Nil =>
-            elseExpr.code shouldBe "else"
-            elseExpr.astChildren.isBlock.astChildren.isCall.name.l shouldBe List("bar")
-        }
-    }
-  }
-
-  "test 14" in {
-    val cpg = code("""
-        |fn main(x: i32, y: i32) {
-        | while x < y {
-        |  foo();
-        | }
-        |}
-        |""".stripMargin)
-
-    inside(cpg.method.name("main").controlStructure.controlStructureTypeExact(ControlStructureTypes.WHILE).l) {
-      case (whileExpr: ControlStructure) :: Nil =>
-        whileExpr.code shouldBe "while x < y {\n  foo();\n }"
-
-        inside(whileExpr.condition.isCall.l) { case (condition: Call) :: Nil =>
-          condition.code shouldBe "x < y"
-          condition.name shouldBe Operators.lessThan
-          condition.methodFullName shouldBe Operators.lessThan
-
-          inside(condition.argument.l) { case (lhs: Identifier) :: (rhs: Identifier) :: Nil =>
-            lhs.name shouldBe "x"
-            rhs.name shouldBe "y"
-          }
-        }
-
-        inside(whileExpr.astChildren.isBlock.l) { case (body: Block) :: Nil =>
-          body.astChildren.isCall.name.l shouldBe List("foo")
-        }
-    }
-  }
-
-  "test 15" in {
-    val cpg = code("""
-        |fn main() {
-        | loop {
-        |  foo();
-        |  break;
-        | }
-        |}
-        |""".stripMargin)
-
-    inside(cpg.method.name("main").controlStructure.controlStructureTypeExact(ControlStructureTypes.WHILE).l) {
-      case (loopExpr: ControlStructure) :: Nil =>
-        loopExpr.code shouldBe "loop {\n  foo();\n  break;\n }"
-
-        inside(loopExpr.condition.isLiteral.l) { case (condition: Literal) :: Nil =>
-          condition.code shouldBe "true"
-          condition.typeFullName shouldBe "bool"
-        }
-
-        inside(loopExpr.astChildren.isBlock.l) { case (body: Block) :: Nil =>
-          body.astChildren.isCall.name.l shouldBe List("foo")
-          body.astChildren.isControlStructure
-            .controlStructureTypeExact(ControlStructureTypes.BREAK)
-            .code
-            .l shouldBe List("break")
-        }
-    }
-  }
-
-  "test 16" in {
-    val cpg = code("""
-        |fn main(x: i32) {
-        | while x < 10 {
-        |  if x == 5 {
-        |   continue;
-        |  }
-        |  break 1;
-        | }
-        |}
-        |""".stripMargin)
-
-    inside(cpg.method.name("main").controlStructure.controlStructureTypeExact(ControlStructureTypes.CONTINUE).l) {
-      case (continueExpr: ControlStructure) :: Nil =>
-        continueExpr.code shouldBe "continue"
-        continueExpr.controlStructureType shouldBe ControlStructureTypes.CONTINUE
-    }
-
-    inside(cpg.method.name("main").controlStructure.controlStructureTypeExact(ControlStructureTypes.BREAK).l) {
-      case (breakExpr: ControlStructure) :: Nil =>
-        breakExpr.code shouldBe "break 1"
-        breakExpr.controlStructureType shouldBe ControlStructureTypes.BREAK
-    }
-  }
-
   "test 17" in {
     val cpg = code("""
         |fn main() {
@@ -579,30 +434,6 @@ class SimpleAstCreationPassTest extends Rust2CpgSuite {
         p.argumentIndex shouldBe 1
       // TODO p.typeFullName shouldBe
       }
-    }
-  }
-
-  "test 33" in {
-    val cpg = code("""
-        |fn main(b: bool) {
-        | if !b {
-        |  foo();
-        | }
-        |}
-        |""".stripMargin)
-
-    inside(cpg.controlStructure.controlStructureTypeExact(ControlStructureTypes.IF).condition.isCall.l) {
-      case (condition: Call) :: Nil =>
-        condition.code shouldBe "!b"
-        condition.name shouldBe Operators.logicalNot
-        condition.methodFullName shouldBe Operators.logicalNot
-        // TODO condition.typeFullName shouldBe "bool"
-
-        inside(condition.argument.l) { case (b: Identifier) :: Nil =>
-          // TODO b.typeFullName shouldBe "bool"
-          b.code shouldBe "b"
-          b.name shouldBe b.code
-        }
     }
   }
